# backend/docs/swagger.yaml
basePath: /
definitions:
  main.DeleteResponse:
    properties:
      filename:
        type: string
      message:
        type: string
    type: object
  main.EditResponse:
    properties:
      filename:
        type: string
      message:
        type: string
    type: object
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.HealthResponse:
    properties:
      status:
        type: string
      time:
        type: string
    type: object
  main.UploadResponse:
    properties:
      filename:
        type: string
      message:
        type: string
    type: object
host: localhost:1234
info:
  contact: {}
  description: Backend for Markdown-editor
  title: Markdown backend
  version: "1.0"
paths:
  /api/file/{filename}:
    delete:
      description: Delete file from server
      parameters:
      - description: Filename to delete
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete response
          schema:
            $ref: '#/definitions/main.DeleteResponse'
        "400":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - CookieAccessToken: []
      summary: Delete file
      tags:
      - files
    get:
      description: Download a file by filename
      parameters:
      - description: Filename to download
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - CookieAccessToken: []
      summary: Download file
      tags:
      - files
    post:
      description: Upload new file to server
      parameters:
      - description: Filename to save
        in: path
        name: filename
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Upload response
          schema:
            $ref: '#/definitions/main.UploadResponse'
        "400":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - CookieAccessToken: []
      summary: Upload file
      tags:
      - files
    put:
      description: Send edited file to server
      parameters:
      - description: Filename to save
        in: path
        name: filename
        required: true
        type: string
      - description: File to save
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Edit response
          schema:
            $ref: '#/definitions/main.EditResponse'
        "400":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - CookieAccessToken: []
      summary: Edit file
      tags:
      - files
  /api/files:
    get:
      description: Get all user files from server
      produces:
      - application/json
      responses:
        "200":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "400":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Error response
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - CookieAccessToken: []
      summary: User files
      tags:
      - files
  /health:
    get:
      description: Check if server respond
      produces:
      - application/json
      responses:
        "200":
          description: Server health status
          schema:
            $ref: '#/definitions/main.HealthResponse'
      summary: Check server health
      tags:
      - health
securityDefinitions:
  CookieAccessToken:
    description: HTTP Only Cookie access_token для авторизации
    in: header
    name: access_token
    type: apiKey
swagger: "2.0"